Extract tray
