+++ app/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
 }
 
 android {
@@ -52,8 +51,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation "com.github.bumptech.glide:okhttp3-integration:4.12.0"
-    kapt 'com.github.bumptech.glide:compiler:4.12.0'
 
     implementation project(path: ':openai')
+    implementation project(path: ':tray')
 }
