--- app/src/main/java/com/mardillu/openai/message/TrayType.kt
@@ -1,10 +1,0 @@
-package com.mardillu.openai.message
-
-/**
- * Created by Hannes Achleitner on 03.11.23.
- */
-enum class TrayType(i: Int) {
-    AI(0),
-    USER(1),
-    ERROR(2)
-}
