+++ settings.gradle
@@ -16,3 +16,4 @@ dependencyResolutionManagement {
 rootProject.name = "ChatGPTWrapper"
 include ':app'
 include ':openai'
+include ':tray'
