+++ tray/src/main/java/com/hannes/tray/message/TrayType.kt
@@ -1,0 +1,8 @@
+package com.hannes.tray.message
+
+
+enum class TrayType(i: Int) {
+    AI(0),
+    USER(1),
+    ERROR(2)
+}
