+++ tray/src/main/res/anim/item_enter.xml
