+++ tray/src/main/res/drawable/message_bubble_bg.xml
