+++ tray/src/main/res/layout/ai_msg_error_view.xml
