+++ tray/src/main/res/layout/ai_msg_view.xml
