+++ tray/src/main/res/layout/user_msg_view.xml
