+++ tray/src/main/res/values/ints.xml
