+++ tray/src/main/res/values/styles.xml
